Add -include-pag-icfg-dump option to control PAG info in ICFG dump
This change puts the dumping of program assignment graph (PAG) information
as part of dumping the interprocedural control-flow graph (ICFG)
under the control of an option. The option defaults to true so as
to preserve the current behavior. When the option is selected,
the PAG pointer is obtained, if the PAG has been constructed,
so the PAG information is added. If the PAG has not been constructed,
the extra PAG information will not be present.