Allow arbitrary ExtAPI.json
Currently, if the default ExtAPI.json path exists, this file will always
be read (irrespective if the user wants to use another path).

This patch allows the user to optionally specify a path to an arbitary
ExtAPI.json file programatically. If this path is specified, it will be
used. Otherwise, the existing lookup process will be used.

Additionally, this patch also fixes a bug in the old ExtAPI::get_FunJson
function: if the default project path was not found, then the SVF_DIR
env variable or NPM is used. However, the jsonFilePath variable was not
updated, so the wrong file would be passed to stat to get the file size.